refactor(recombination): split HMM detection into focused submodules
- Separate the single detection file into submodules by concern (parameters, observations, Viterbi decoding, forward-backward scoring, config, result), so each concern is independently readable and testable instead of interleaved in one file.
- Group the tree-based parameter estimator and the co-located tests under the same `recombination/` directory as its family rather than as loose analyze-level siblings.
